requirements.txt
# 精准动态教辅 - Python后端依赖 (简化版本)

# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0

# 数据库
sqlalchemy==2.0.23
alembic==1.12.1

# AI模型集成
openai==1.3.7
requests==2.31.0

# 数据验证
pydantic==2.5.0

# 密码加密
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# 文件上传处理
python-multipart==0.0.6

# 环境变量管理
python-dotenv==1.0.0

# 日期时间处理
python-dateutil==2.8.2

# JSON处理
orjson==3.9.10

# 测试框架
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# 图像处理 (可选，后续添加)
# pillow>=10.0.0
# opencv-python>=4.8.0

# PDF生成 (可选，后续添加)
# reportlab>=4.0.0
# weasyprint>=60.0.0